Fix for Helm on CoreOs using Docker
Helm (tiller) requires socat to run. CoreOS doesnt supply socat
with their OS (See https://github.com/coreos/bugs/issues/1114) hence
we have to add it to the system during Kubernetes Deployment.
While there is a solution for this issue when running Kubernetes on CoreOS with Rkt
(Kubelet Wrapper https://coreos.com/kubernetes/docs/latest/kubelet-wrapper.html)
there is no when running Kubernetes on CoreOS with
Docker. For this reason we compiled a static socat binary we copy
to /opt/bin or another directory and add the PATH to the Kubelet Docker
Container to run Helm (tiller) on the Kubernetes Cluster.